requirements.txt
GitPython==2.1.14
gitdb==0.6.4
attrs==19.1.0
smmap==0.9.0